The creation of the main weapon/tool and the item pickup functions
The "GG3000" script is the code for our main weapon/tool, the way it works is it firstly looks for a collided object which is found from another collider, which is why we call collided position in the awake. After that, we take the rigidbody of the desired object and make it kinematic, in the case of the sucking function, the object moves its position towards the player, until the distance is low enough for the object to be destroyed and saved as a variable count, while in the move/create function, the object's transform position is given the data of the transform position of an empty game object we have in unity name "ObjectHolder" until input is given for the function to end.
The "PickUp" script is very simple, it checks for a trigger enter, and checks the tag of the object collided with, if the object has the right tag, after input is given we destroy the object and we store a variable charge.